requirements.txt
Django
djangorestframework
# drf-yasg
drf-spectacular